standalone (non-NVFlare) version of minimal training to have an example what the difference is
